yub20231006-final

+++ BookRecommendService/src/book/test/FavoriteDAO.java
@@ -102,7 +102,7 @@ public class FavoriteDAO {
 	private void delete(FavoriteDTO input) throws SQLException {
 		Connection conn = pool.getConnection();
 		Statement stmt = conn.createStatement();
-		String sql = "delete from favorites where bookid=" + input.getBookID();
+		String sql = "delete from favorites where bookid='" + input.getBookID() + "' and userID = '" + input.getUserID() + "'";		
 		int result = stmt.executeUpdate(sql);
 		
 		stmt.close();
